Enable autosave to localStorage (if available)
This is a bit of a mess at the moment, because of whitespace
problems and some duplicated code. Will get that cleaned up.

The way it works is relatively simple:

When starting an edit, start up an interval that writes the current
data to localStorage every 10 seconds.

When the save is done, remove the backup.

If when you startup, there is a backup, ask if it should be used.

Just realized the timeout is not being cancelled, so will add that.