Delete currentFields after a 404
Otherwise if there was a previous tiddler around, it's etag
can mess up editing a new one. See:

  https://github.com/TiddlySpace/tiddlyspace/issues/1014

This is an example of the various bits of global state that
ought to be managed in a more effective way but aren't because
of the YAGNI approach being used to the development of the tool.
As long as things get fixed, that's fine and good.