support markdown extraclusion
fix #18

In the process make it so tiddlers which are tiddlywikitext
are text/x-tiddlywiki instead of no type. This is good. This
is the future. Required a change over here:

   https://github.com/tiddlyweb/tiddlywebwiki/commit/7e90035df26d43b2dbf7805535c59171553e87fe

The same syntax is used to create an extraclusion, but the output
on save is different depending on markdown or tiddlywikitext.