fix #28 by implementing a different guard on privacy toggle event
The code should only bind the event if the current tiddler has
never been saved (that is, the tiddler is new) because it doesn't
(yet) do a PUT + DELETE combo for a move from public to private;
instead it just sets the target bag for a new tiddler.

In the old code `currentBag` was being used as the guard. Prior to
ebd19f51eb4cb556e89119404eed48acf12b96dc this was okay, but in that
commit currentBag is set from the start.

Now a new variable `freshTiddler` is used to indicate when we have
a branch new tiddler.

Sadly this was an easy enough fix that the planned rewrite is put
off yet again.